codex: split exec-server smoke into child PR (#28120)
