+++ codex-rs/core/tests/remote_env_windows/BUILD.bazel
@@ -2,23 +2,20 @@ load("//bazel/rules/testing:wine.bzl", "wine_rust_test")
 
 wine_rust_test(
     name = "smoke-test",
-    timeout = "moderate",
-    srcs = [
-        "non_native_cwd_tests.rs",
-        "remote_env_windows_test.rs",
-    ],
+    timeout = "short",
+    srcs = ["remote_env_windows_test.rs"],
     crate_name = "remote_env_windows_test",
     crate_root = "remote_env_windows_test.rs",
     windows_binaries = {
         "wine-windows-exec-server": "//codex-rs/exec-server/testing:windows-exec-server",
     },
     deps = [
+        "//bazel/rules/testing/wine:wine_test_support",
         "//codex-rs/core/tests/common",
         "//codex-rs/exec-server",
-        "//codex-rs/exec-server/testing:wine-exec-server-harness",
         "//codex-rs/features",
         "//codex-rs/protocol",
-        "//codex-rs/utils/path-uri",
+        "//codex-rs/utils/cargo-bin",
         "@crates//:anyhow",
         "@crates//:pretty_assertions",
         "@crates//:serde_json",
