+++ codex-rs/exec-server/testing/BUILD.bazel
@@ -1,21 +1,4 @@
-load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_library")
-
-rust_library(
-    name = "wine-exec-server-harness",
-    testonly = True,
-    srcs = ["wine_exec_server_harness.rs"],
-    crate_name = "wine_exec_server_harness",
-    target_compatible_with = ["@platforms//os:linux"],
-    visibility = ["//codex-rs/core/tests/remote_env_windows:__pkg__"],
-    deps = [
-        "//bazel/rules/testing/wine:wine_test_support",
-        "//codex-rs/utils/cargo-bin",
-        "//codex-rs/utils/pty",
-        "@crates//:anyhow",
-        "@crates//:tempfile",
-        "@crates//:tokio",
-    ],
-)
+load("@rules_rust//rust:defs.bzl", "rust_binary")
 
 rust_binary(
     name = "windows-exec-server",
